dataset_pipeline/requirements.txt
# Parsing
beautifulsoup4>=4.12.0
pdfminer.six>=20221105
python-docx>=1.1.0

# VLM (Qwen2-VL OCR)
transformers>=4.40.0
torch>=2.0.0
pillow>=10.0.0

# Utils
tqdm>=4.66.0
qwen-vl-utils
pydantic
torch 
torchvision 
torchaudio --index-url https://download.pytorch.org/whl/cu121  # для CUDA 12.x
datasets
hf_transfer
tensorboard